feat: Streamlined gulp process, added sass map
Added a SASS map file, which makes debugging in the browser. Now, when you want to trace a CSS line to source, you can, as the map file should map each CSS line back to its original SASS location and file.

Also took the opportunity here to simplify the gulpfile, as it was unnecessarily complex due to legacy code still being present. Reduced the dependencies by two in the npm package, as a result.